+++ glfw/posix_time.h
@@ -1,5 +1,5 @@
 //========================================================================
-// GLFW 3.3 POSIX - www.glfw.org
+// GLFW 3.4 POSIX - www.glfw.org
 //------------------------------------------------------------------------
 // Copyright (c) 2002-2006 Marcus Geelnard
 // Copyright (c) 2006-2017 Camilla Löwy <elmindreda@glfw.org>
@@ -34,7 +34,7 @@
 //
 typedef struct _GLFWtimerPOSIX
 {
-    bool    monotonic;
+    bool        monotonic;
     uint64_t    frequency;
 
 } _GLFWtimerPOSIX;